Module sp_fragnova::fragnova

A runtime interface for the Fragnova Blockchain

Background:

#[sp_runtime_interface::runtime_interface] is an attribute macro for transforming a trait declaration into a runtime interface.

A runtime interface is a fixed interface between a Substrate compatible runtime and the native node. This interface is callable from a native and a wasm runtime. The macro will generate the corresponding code for the native implementation and the code for calling from the wasm side to the native implementation. The macro expects the runtime interface declaration as trait declaration.

Source: https://paritytech.github.io/substrate/latest/sp_runtime_interface/attr.runtime_interface.html

Structs

HostFunctions

Provides implementations for the extern host functions.

Functions

get_geth_url

Get the URL of the Fragnova-owned Geth Node

on_new_fragment

TODO

say_hello_world

A function that can be called from native/wasm.