Module sp_fragnova::protos

Types that will be used by the Protos pallet Types that will be used in ths Protos Pallet

Structs

Proto

Struct of a Proto-Fragment

ProtoPatch

Struct of a Proto-Fragment Patch

Enums

LinkSource

TODO: Documentation

LinkedAsset

Types of Assets that are linked to a Proto-Fragment (e.g an ERC-721 Contract etc.)

ProtoData

Enum that indicates how a Proto-Fragment data can be fetched

ProtoOwner

Types of Proto-Fragment Owners

UsageLicense

Enum that indicates how a Proto-Fragment can be used